requirements.txt
pypdf>=3.0.0
llama-cpp-python>=0.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
beautifulsoup4>=4.12.0
ddgs>=9.0.0

# Para usar Ollama (alternativa con GPU automática)
# No requiere pip install, solo tener Ollama instalado
# Descargar desde: https://ollama.com/download
